+++ HISTORY.rst
@@ -87,6 +87,7 @@
   instead of during rendering.
 - Split Bootstrap specific markup and style into the `appearance_bootstrap`
   app.
+- Split document indexing serializers into separate modules.
 
 4.7.1 (2024-06-04)
 ==================